- remove store/app-store.js in favor of store/app-store/index.js - moved reducers, actions, selectors into their own files - update .css to be valid - update DynamicToolbarButton
