+++ plugins/third-party-plugin/src/components/ShareDocumentButton/ShareDocumentButton.jsx
@@ -3,7 +3,6 @@ import React from 'react';
 import { Icon, external } from '@wordpress/icons';
 import { VisuallyHidden } from '@wordpress/components';
 import { useEditorContext } from '@graphiql/react';
-import LZString from 'lz-string';
 import { useCopyToClipboard } from '../../../../../src/hooks/useCopyToClipboard';
 
 /**
@@ -51,22 +50,4 @@ export const ShareDocumentButton = ({ ToolbarButton }) => {
 	);
 };
 
-/**
- * Compresses and encodes a query parameter object for use in a shareable URL.
- *
- * @param {Object} obj The object containing query parameters to be compressed and encoded.
- * @return {string} A compressed and encoded string representing the query parameters.
- */
-export function getHashedQueryParams( obj ) {
-	if ( typeof obj !== 'object' || obj === null ) {
-		console.error( 'Input must be a non-null object' );
-		return '';
-	}
-	try {
-		const queryParamString = JSON.stringify( obj );
-		return LZString.compressToEncodedURIComponent( queryParamString );
-	} catch ( error ) {
-		console.error( 'Failed to compress query parameter object:', error );
-		return '';
-	}
-}
+
