+++ src/store/editor-toolbar/actions.js
@@ -1,0 +1,9 @@
+const actions = {
+	registerButton: ( name, component ) => ( {
+		type: 'REGISTER_BUTTON',
+		name,
+		component,
+	} ),
+};
+
+export default actions;
