+++ src/store/editor-toolbar/selectors.js
@@ -1,0 +1,5 @@
+const selectors = {
+	buttons: ( state ) => state.buttons,
+};
+
+export default selectors;
